stricter PR input
Do not accept pull request URLs that have trailing information.

With this change, we now bail if the user tries to provide a specific
commit in a PR, like so:
https://github.com/ceph/ceph/pull/10584/commits/af3d3e25823ca9b40ddf09835edb82795ac68f33

For now I intend this downstream-cherry-picker app is only designed for
cherry-picking entire PRs, nothing more. By erroring on these longer
URLs, this makes our behavior explicit to the user.