README: avoid "git pull" instruction
"git pull" will merge the local branch with the remote. If I've rebased
the -patches branch to a newer upstream version, "git pull" will create
a mess.

We really want users to do a hard reset to the remote -patches branch,
so they're working from the latest code in the remote.