get_sha_range(): better first and last sha1 detection
Prior to this change, we would simply return the first and last list
elements from the GitHub "commit" API for a PR.

Unfortunately, GitHub uses some strange commit sorting algorithm that I
do not understand. It is hinted at here:
https://help.github.com/articles/why-are-my-commits-in-the-wrong-order/
, and I think the order in which a branch is re-pushed and rebased
during PR has something to do with it.

In one instance (Ceph PR 12917), this caused downstream-cherry-picker to
come up with the incorrect range of commits:

  The GitHub REST API and UI said that (698250) "rgw: s3: secure_port
  should override port" is the final commit in the PR.

  `git fetch https://github.com/ceph/ceph pull/12917/head` says "rgw:
  Handle multiple listening addreses" is the final one.

Use a second API query to properly determine "head" and "base", and then
calculate the "first" and "last" commits from those values. Add tests to
verify that we're now correct in PR 12917's case.