get_sha_range(): handle PR target moving
Prior to this change, if a PR was originally targeted at one branch and
then moved to a second branch, downstream-cherry-picker would fail to
determine the first commit sha1 to cherry-pick.

GitHub's main Pull Request API endpoint populates the "base" value as
the new destination branch's sha1. and it's possible (likely?) that when
a PR is re-targeted like this, none of the commits in the PR will have
the new base sha1 as a parent.

When this happened, downstream-cherry-picker would return "None" for
"first" in get_sha_range(), which causes the "git cherry-pick" argument
to look like "None~..1926adc826de9ddaaa04138ce2a0b276eeda7ad1", which of
course fails.

In these cases, fall back to just choosing the "first" commit in the PR
as the first one to cherry-pick.