get_sha_range(): find the first commit correctly
Previous changes broke the ability to properly find the first commit in
straightforward PRs.

Rather than doing a search for a commit with "base" as the
parent, determine "first" by comparing the parents to the list of
commits. In particular, search for a commit that does not have a parent
in our current set of commits.

Raise RuntimeError if we couldn't find a "first" commit this way, or if
this method results in more than one "first".

Raise NotImplementedError if the PR contains a merge commit, because
git-cherry-pick is going to bail on this range anyway without the "-m"
argument.

Add API fixtures from several more PRs to make sure we have this
behavior down correctly.