silence output from git-cat-file
If git cat-file fails, it prints a "fatal:" error message to stderr.

We expect this to fail occasionally because we're using it as a test to
discover if we need to fetch the PR locally.

Add a "log_output" kwarg to git() to allow us to disable the output from
the subcommand call, and specifically disable this during the git
cat-file operation.

A safer fix would be to look for that specific "fatal:" error message
string after the git() call and filtering *that* specifically, but that
would require another kwarg to git() to optionally disable raising
subprocess.CalledProcessError and return the output instead. I'm not
sure I want to go there yet.